Replace destructor with closeBrowser method
The browser destructor was renamed to closeBrowser. This new method is now properly called at the end of the migration process to explicitly close the browser. Also, the relationship between PageBuilder and Browser classes has been slightly modified, allowing direct access to the browser instance from PageBuilder instance, to facilitate the calling of the closeBrowser method.