Refactor and improve elements and forms in Theme Anthem
The `Element.php` and `Forms/Form.php` classes under `Theme Anthem` in the MBMigration library have been refactored for better code readability. Changes also include the addition of formId to the `wrapperForm` method and the adjustment in the conditions for checking background filename and child node's types. The form referenced when building the wrapper form is now more flexible.