Update theme to use BrowserPHP and ColorConverter
Replaced the deprecated Browser and manually implemented color conversion functionality with BrowserPHP and ColorConverter respectively. This change increases the efficiency of the Anthem theme by enabling the use of PHP features for Browser operations and providing a utility for color conversion. It also offers better maintainability by removing the hard-coded convertColor function.