Refactor MBMigration theme to use BrowserPHP and ColorConverter classes
Updated the Anthem theme under MBMigration to make use of the more efficient BrowserPHP instead of the older Browser for handling browser actions. Separately outsourced the color conversion functionality to a dedicated utility, ColorConverter, replacing the manually defined convertColor function. This improves code maintainability and encapsulation, further increasing the efficiency of the operations.