Update detailPage setting to use encoded placeholder
Reformatted the 'detailPage' setting in Event Grid, Gallery, Calendar, List, and Sermons Layouts. Instead of directly inserting the 'collectionItemsForDetailPage' as the 'id' parameter in 'brizy_dc_url_post', it is now being base64 encoded and used as a placeholder. This update is expected to provide easier read and maintenance while preventing potential string parsing issues.