Add page mapping to theme and head elements
Added the 'pageMapping' variable to various theme elements in Anthem.php and head elements in Head.php. This change allows these elements to have direct access to the page mapping for better navigation handling. The 'pageMapping' data is now included in the array data that is passed to the methods. This was done to effectively manage internal links and improve navigation scheme development.