Add PrayerForm to dynamic elements and globalBlocksKit
A new PrayerForm has been added to the dynamic elements in the Anthem theme structure. This PrayerForm uses DynamicElement as a base class and has been wired up with the corresponding logic and presentation aspects. In addition, this form has been included in the globalBlocksKit.json file to ensure its widespread availability.