Add node related methods and update Anthem class
Added `hasNode` and `getNodeText` methods to the BrowserPagePHP class. These additions provide functionalities to check if a node exists and fetch text from a node, respectively. An update is also made to the Anthem.php class to incorporate these new methods. Also, some unused code is removed and type hint of $elementOptions is updated in DynamicElement.php.