Add .editorconfig and adjust indentations in getRootPropertyStyles
This commit introduces a new .editorconfig file to enforce consistent coding style. It sets the general indent style to spaces with an indent size of 2, except for PHP files which have an indent size of 4. Additionally, the indentations in getRootPropertyStyles.ts file have been adjusted to comply with these settings.