feat(Subpalette): function to DOM package
A new file named detectSubpalette.ts has been added to the Dom package in the elements/src which contains a new function called detectSubpalette. This function helps to detect subpalettes in a given HTML selector. Additionally, it's been imported and added to the object export in the Dom/index.ts file.

--

--

--