Update styles and improve logging strings formatting in Anthem theme
In this update, the title and hover color paths in GridMediaLayout are adjusted. Code related to ListMediaLayout in AnthemElementsController is removed. Furthermore, the string formatting for information and warning logs throughout the Anthem theme has been improved to enhance readability. The code handling 'grid-media-layout' or 'list-media-layout' in Anthem.php is also updated.