Update title color source and navigation type
The commit modifies the `GridMediaLayout.php` file, changing the source of `titleColorHex` and `hoverTitleColorHex` to be from 'palette' instead of 'sermon'. Also, in the `BrowserPHP.php` file, the navigation wait type was updated from 'LOAD' to 'NETWORK_IDLE'. An explicit newline was added at the end of the document as well to adhere to good coding standards.