Add SermonFeatured dynamic element for sermons theme
A new dynamic element, SermonFeatured, has been added to the Anthem theme, allowing for the feature of specific sermons in the section. This offers more customization options for the sermon section of the layout, including styling and filtering features. The corresponding main and head layouts for the SermonFeatured element are added to the globalBlocksKit.json file.