Add slug creation method and refactor codebase
A new method for slug creation in Element.php was added, which aims to create URL-friendly strings. Additionally, redundant uses of conditions and styles in AnthemElementsController.php and Anthem.php, respectively, were refactored. The overall purpose was to improve readability, maintainability, and performance of the codebase. The `devMode` in the settings of the index.php file was also hard-set to 'dev'.