Update element settings related to hover and subscribe event
The settings for elements in the Anthem theme have been updated to include hover effects on various elements and a new subscribe event button. This entails changes to 'showPreviousPage' setting and introduction of 'subscribeEventButton' and 'hoverSubscribeEventButton' settings with respective background color, opacity and palette attributes. Hover effects for title and listItemTitle are also updated.